Fixed Type Mismatch
error[E0308]: mismatched types
  --> src/main.rs:22:57
   |
22 |     .window_mode(conf::WindowMode::default().dimensions(WIDTH, HEIGHT))
   |                                                         ^^^^^ expected f32, found u32
   |
help: you can cast an `u32 to `f32`, producing the floating point representation of the integer,
   |                                              rounded if necessary
22 |     .window_mode(conf::WindowMode::default().dimensions(WIDTH as f32, HEIGHT))
   |                                                         ^^^^^^^^^^^^

error[E0308]: mismatched types
  --> src/main.rs:22:64
   |
22 |     .window_mode(conf::WindowMode::default().dimensions(WIDTH, HEIGHT))
   |                                                                ^^^^^^ expected f32, found u32
   |
help: you can cast an `u32 to `f32`, producing the floating point representation of the integer,
   |                                              rounded if necessary
22 |     .window_mode(conf::WindowMode::default().dimensions(WIDTH, HEIGHT as f32))
   |                                                                ^^^^^^^^^^^^^

+++ src/main.rs
@@ -11,8 +11,8 @@ use ggez::timer::check_update_time;
 
 use rand::Rng;
 
-const WIDTH: u32 = 320;
-const HEIGHT: u32 = 200;
+const WIDTH: f32 = 320.0;
+const HEIGHT: f32 = 200.0;
 const STARS_COUNT: u16 = 200;
 const XY_RANGE: i32 = 25;
 const MAX_DEPTH: u32 = 32;
